ındex

+++ FinallyProjectUI/Controllers/ProductController.cs
@@ -151,29 +151,74 @@ namespace FinallyProjectUI.Controllers
             return RedirectToAction("Index", "Product");
         }
 
-        [HttpGet]
+        [HttpDelete]
         public IActionResult Delete(int Id)
         {
-            var productToDelete = _context.Products.FirstOrDefault(x => x.Id == Id);
-            var ImagesTodelete = _context.PImages.Where(u => u.ProductId == Id).Select(u => u.ImageUrl);
-            foreach(var image in ImagesTodelete)
+            if(Id != 0)
             {
-                string imageUrl = "Images\\" + image;
-                var toDelteImageFromFolder = Path.Combine(_HostEnvironment.WebRootPath, imageUrl.TrimStart('\\'));
-                DeleteAImage(toDelteImageFromFolder);
+                var productToDelete = _context.Products.FirstOrDefault(x => x.Id == Id);
+                var ImagesTodelete = _context.PImages.Where(u => u.ProductId == Id).Select(u => u.ImageUrl);
+                foreach (var image in ImagesTodelete)
+                {
+                    string imageUrl = "Images\\" + image;
+                    var toDelteImageFromFolder = Path.Combine(_HostEnvironment.WebRootPath, imageUrl.TrimStart('\\'));
+                    object value = DeleteAImage(toDelteImageFromFolder);
+                }
+                if (productToDelete.HomeImageUrl != "")
+                {
+                    string imageUrl = "Images\\" + productToDelete.HomeImageUrl;
+                    var toDelteImageFromFolder = Path.Combine(_HostEnvironment.WebRootPath, imageUrl.TrimStart('\\'));
+                    DeleteAImage(toDelteImageFromFolder);
+                }
+                _context.Products.Remove(productToDelete);
+                _context.SaveChanges();
             }
-            if(productToDelete.HomeImageUrl != "")
+            else
             {
-                string imageUrl = "Images\\" + productToDelete.HomeImageUrl;
-                var toDelteImageFromFolder = Path.Combine(_HostEnvironment.WebRootPath, imageUrl.TrimStart('\\'));
-                DeleteAImage(toDelteImageFromFolder);
+                return Json(new {success = false, message= "Öge silinemedi"});
             }
-            _context.Products.Remove(productToDelete);
-            _context.SaveChanges();
-            return RedirectToAction("Index", "Product");
+
+
+            return Json(new { success = true, message = "Silme başarılı" });
         }
 
-        private void DeleteAImage(string toDelteImageFromFolder)
+        public IActionResult DeleteAImage(string Id)
+        {
+            int routeId = 0;
+            if (Id !=null)
+            {
+                if(!Id.Contains("Home"))
+                {
+                    var ImageToDeleteFromPImage= _context.PImages.FirstOrDefault(u => u.ImageUrl == Id);
+                    if (ImageToDeleteFromPImage!=null) 
+                    {
+                       routeId = ImageToDeleteFromPImage.ProductId;
+                        _context.PImages.Remove(ImageToDeleteFromPImage);
+                    }
+                }
+                else
+                {
+                    var ImageToDeleteFromProduct = _context.Products.FirstOrDefault(u => u.HomeImageUrl == Id);
+                    if(ImageToDeleteFromProduct!=null) 
+                    {
+                      ImageToDeleteFromProduct.HomeImageUrl = "";
+                        routeId = ImageToDeleteFromProduct.Id;
+                        _context.Products.Update(ImageToDeleteFromProduct);
+                    }
+                }
+                string ImageUrl = "Images\\" + Id;
+                var toDeleteImageFromFolder =Path.Combine(_HostEnvironment.WebRootPath, ImageUrl);
+                object value = DeleteAImage(toDeleteImageFromFolder);
+                _context.SaveChanges();
+                return Json(new {success = true, message="Resim başarıyla silindi", id= routeId});
+            }
+            return Json(new { success = false, message = "resim silinemedi" });
+           
+        }
+
+
+
+        private static void DeleteAImage(string toDelteImageFromFolder)
         {
             if (System.IO.File.Exists(toDelteImageFromFolder))
             {
